Refactor and restructure utils module in reports app
- Moved utility files to `utils` subdirectory for better organization.
- Adjusted imports accordingly across views and signals.
- Updated file paths for generated report files in `DeliveryReport` model.
- Enhanced app readiness by linking signals in the `apps.py`.
