+++ backend/reports/utils/image_utils.py
