+++ backend/reports/utils/pdf_utils.py
