+++ backend/reports/utils/plate_recognition_utils.py
